feat(devex): Introduce prettier-plugin-sort-imports to sort our imports
We've recently moved from eslint to oxlint which caused us to lose our lints that would sort our imports. Let's introduce a new prettier plugin (prettier-plugin-sort-imports) to compensate for that. This plugin is highly customizable, and I tried to make sure the rules it's following align closely to what we had before.